trace: Put LogLevel, MessageTemplate by default
includeLogLevelAsProperty, includeMessageTemplateAsProperty bools
for trace were true before and were changed (accidentally, perhaps)
to false in commit 1eb4fb4 ("Trace does not include LogLevel and
MessageTemplate by default (#36)", 2019-01-28). So revert those to be
true so that highest version in 2.x has the intended behavior which
is same as v3.x.